fix(registration): preserve sub-pixel shift precision (remove int-cast)
Both backends computed the shift to upsample_factor=10 (0.1 px) then stored
int(np.round(shift*df)), discarding the sub-pixel part. On the fixture every
recovered offset was a literal integer and the worst pairwise error was 0.52 px
(pair 2,3) -- the likely source of seam ghosting / duplicated edge features.

Store the float shift in register_pair_worker and register_pairs_readahead;
pairwise_metrics type hint int->float. Measured on the fixture: worst pairwise
0.52 -> 0.085 px, worst global 0.41 -> 0.065 px (6x, now at the 0.1 px grid).

Flip the strict-xfail test_accuracy_pairwise to a passing sub-pixel gate
(atol 0.15 px, sound bound over the measured 0.085); tighten the global gate
0.5 -> 0.15. Add a register_pair_worker unit test (the batched path is otherwise
unexercised by the fixture). Re-capture golden_metrics.json (sub-pixel floats,
intentional); loosen the regression atol 1e-6 -> 1e-4 for FFT-float stability.
Full suite 201 passed, 0 xfailed.

+++ src/tilefusion/core.py
@@ -254,7 +254,7 @@ class TileFusion:
     def _init_pipeline_state(self) -> None:
         """Initialize the mutable registration/fusion pipeline state."""
         # State
-        self.pairwise_metrics: Dict[Tuple[int, int], Tuple[int, int, float]] = {}
+        self.pairwise_metrics: Dict[Tuple[int, int], Tuple[float, float, float]] = {}
         self.global_offsets: Optional[np.ndarray] = None
         self.offset: Optional[Tuple[float, float]] = None
         self.unpadded_shape: Optional[Tuple[int, int]] = None
